Extract span scanner
